For #93: Fix percentage calculation

+++ app/Helpers/MenuHelper.php
@@ -11,9 +11,9 @@ class MenuHelper
      * @param int $disks Number of disks in the menu set
      * @param int $missing Number of missing disks
      *
-     * @return int Percentage of completion of a menu set
+     * @return float Percentage of completion of a menu set
      */
-    public static function percentComplete(int $disks, int $missing): int
+    public static function percentComplete(int $disks, int $missing): float
     {
         return ($disks - $missing) / $disks * 100;
     }
